requirements.txt
g2pK==0.9.4
gensim==4.2.0
hangul-jamo==1.0.1
inko-py==1.0.0
wget==3.2
python-mecab-ko==1.0.14